# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "Backend routes for CSElectives 2.0",
		"version": "1.2.0",
		"title": "CSElectives API Routes",
		"contact": {
			"email": "projects.cselectives@csesoc.org.au"
		},
		"license": {
			"name": "CSElectives 2.0",
			"url": "https://github.com/csesoc/cselectives-v2/blob/develop/LICENSE"
		}
	},
	"paths": {
...
    "/allCourses": {
			"get": {
				"tags": ["Courses"],
				"summary": "Gets a list of all course objects",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Course"
									}
								}
							}
						}
					},
					"500": {
						"description": "An internal server error occurred"
					}
				}
			}
		},
		"/courses": {
			"get": {
				"tags": ["Courses"],
				"summary": "Gets a list of all course objects per 25 offset",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Course"
									}
								}
							}
						}
					},
					"500": {
						"description": "An internal server error occurred"
					}
				}
			}
		},
		"/courses/${courseCode}": {
			"put": {
				"tags": ["Courses"],
				"summary": "Puts a course with new courseCode into database",
				"parameters": [
					{
						"in": "path",
						"name": "courseCode",
						"required": true,
						"schema": {
							"type": "string",
							"minimum": 1
						},
						"description": "The course code"
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Course"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Course"
								}
							}
						}
					},
					"400": {
						"description": "Bad request"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"Course": {
				"type": "object",
				"properties": {
					"courseCode": {
						"type": "string"
					},
					"archived": {
						"type": "boolean"
					},
					"attributes": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"calendar": {
						"type": "string"
					},
					"campus": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"enrolmentRules": {
						"type": "string"
					},
					"equivalents": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Course/properties/courseCode"
						}
					},
					"exclusions": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Course/properties/courseCode"
						}
					},
					"faculty": {
						"type": "string"
					},
					"fieldOfEducation": {
						"type": "string"
					},
					"genEd": {
						"type": "boolean"
					},
					"level": {
						"type": "integer"
					},
					"school": {
						"type": "string"
					},
					"studyLevel": {
						"type": "string"
					},
					"terms": {
						"type": "array",
						"items": {
							"type": "integer"
						}
					},
					"title": {
						"type": "string"
					},
					"uoc": {
						"type": "integer"
					},
					"rating": {
						"type": "number"
					},
					"reviewCount": {
						"type": "integer"
					},
					"overallRating": {
						"$ref": "#/components/schemas/Review/properties/overallRating"
					},
					"manageability": {
						"$ref": "#/components/schemas/Review/properties/manageability"
					},
					"usefulness": {
						"$ref": "#/components/schemas/Review/properties/usefulness"
					},
					"enjoyability": {
						"$ref": "#/components/schemas/Review/properties/enjoyability"
					}
				}
			},
...
		}
	}
}